refactor: update dependencies and example app
Previously, RNSVG was incorrectly used as dependency instead of devDependency.
Furthermore, the version was "^12.1.0". RNSVG has introduced fabric support
from version 13.x.x.
Since this package doesn't have to rely on specific version of RNSVG, it's better
to have this change.

The old example is removed and new example with RN 0.69.5 is added.
Yarn is used as package manager. This is a common convention for React apps.